implemented acton taken revised blocker

+++ pages/api/chairperson_suggestion_item_action_taken/[id]/index.ts
@@ -31,6 +31,46 @@ export default async function handler(
       await validator.validate(req.query);
 
       const { id } = validator.cast(req.query);
+
+      const departmentRevision = await prisma.departmentRevision.findFirst({
+        where: {
+          IMFile: {
+            DepartmentReview: {
+              ChairpersonReview: {
+                ChairpersonSuggestion: {
+                  ChairpersonSuggestionItem: {
+                    some: {
+                      ChairpersonSuggestionItemActionTaken: {
+                        id: {
+                          equals: id,
+                        },
+                      },
+                    },
+                  },
+                },
+              },
+            },
+          },
+          OR: [
+            {
+              ReturnedDepartmentRevision: {
+                is: null,
+              },
+            },
+            {
+              ReturnedDepartmentRevision: {
+                SubmittedReturnedDepartmentRevision: {
+                  is: null,
+                },
+              },
+            },
+          ],
+        },
+      });
+      if (departmentRevision) {
+        throw new Error("IM already revised.");
+      }
+
       const chairpersonSuggestionItemActionTaken =
         await prisma.chairpersonSuggestionItemActionTaken.findFirstOrThrow({
           where: {
@@ -68,6 +108,45 @@ export default async function handler(
       );
 
       const { id } = validator.cast(req.query);
+
+      const departmentRevision = await prisma.departmentRevision.findFirst({
+        where: {
+          IMFile: {
+            DepartmentReview: {
+              ChairpersonReview: {
+                ChairpersonSuggestion: {
+                  ChairpersonSuggestionItem: {
+                    some: {
+                      ChairpersonSuggestionItemActionTaken: {
+                        id: {
+                          equals: id,
+                        },
+                      },
+                    },
+                  },
+                },
+              },
+            },
+          },
+          OR: [
+            {
+              ReturnedDepartmentRevision: {
+                is: null,
+              },
+            },
+            {
+              ReturnedDepartmentRevision: {
+                SubmittedReturnedDepartmentRevision: {
+                  is: null,
+                },
+              },
+            },
+          ],
+        },
+      });
+      if (departmentRevision) {
+        throw new Error("IM already revised.");
+      }
 
       const chairpersonSuggestionItemActionTaken =
         await prisma.chairpersonSuggestionItemActionTaken.delete({
