+++ pages/api/content_editor_suggestion_item_action_taken/[id]/index.ts
@@ -69,6 +69,47 @@ export default async function handler(
 
       const { id } = validator.cast(req.query);
 
+      const iMERCCITLRevision = await prisma.iMERCCITLRevision.findFirst({
+        where: {
+          IMFile: {
+            IMERCCITLRevision: {
+              IMERCCITLReviewed: {
+                SubmittedContentEditorSuggestion: {
+                  ContentEditorSuggestion: {
+                    ContentEditorSuggestionItem: {
+                      some: {
+                        ContentEditorSuggestionItemActionTaken: {
+                          id: {
+                            equals: id,
+                          },
+                        },
+                      },
+                    },
+                  },
+                },
+              },
+            },
+          },
+          OR: [
+            {
+              ReturnedIMERCCITLRevision: {
+                is: null,
+              },
+            },
+            {
+              ReturnedIMERCCITLRevision: {
+                SubmittedReturnedIMERCCITLRevision: {
+                  is: null,
+                },
+              },
+            },
+          ],
+        },
+      });
+      if (iMERCCITLRevision) {
+        throw new Error("IM already revised.");
+      }
+      
       const contentEditorSuggestionItemActionTaken =
         await prisma.contentEditorSuggestionItemActionTaken.delete({
           where: {
